Force strictness for CodeBlock's `mappend`.
CodeBlock is the monoid for the Ivory monad's WriterT. MonadLib's
WriterT is strict in the output monoid, but this particular monoid was
insufficiently strict for good performance. Code running in the Ivory
monad for a large number of (>>=) operations could build up a pile of
CodeBlock thunks as a result.