Added ProcPtr to IvoryStore, enabling 'store' to work on ProcPtr.
This commit message is Literate Haskell to explain the commit's purpose.

> {-# LANGUAGE TypeOperators #-}
> {-# LANGUAGE DataKinds #-}
> import Ivory.Language
> import Ivory.Compile.C.CmdlineFrontend
> m = package "StoreProcPtr" $ do
>     incl setFnExample
>     defMemArea fnExtn
> main = runCompiler [m] [] initialOpts

The specific use case that brought this to my attention was the need to
interface Ivory with some APIs that relied on callbacks.  Due to some other
limitations, I had to write by hand C wrappers for these callbacks.

However, I wanted Ivory to have the ability to still set this callback at
runtime somehow, and so I had the C wrapper make a call to a function pointer,
and I tried to set this function pointer's value from within Ivory.

Suppose that 'fnExtn' below is that function pointer.

> type Fn = '[Uint16] :-> ()
> fnExtn :: MemArea (Stored (ProcPtr Fn))
> fnExtn = area "fnExtn" Nothing

I have an Ivory procedure 'fnExample' that I would like to assign to 'fnExtn'
sometime at runtime:

> fnExample :: Def Fn
> fnExample = proc "fnExample" $ \ _ -> body $ retVoid

And so I also have a function 'setFn' which, given a procedure name and
another Ivory procedure, generates an Ivory procedure that sets that global
function pointer 'fnExtn' to that first procedure.

> setFn :: String -> Def Fn -> Def('[] :-> ())
> setFn name fn = proc name $ body $
>                      store (addrOf fnExtn) (procPtr fn)

...which I've instantiated thus with fnExample as 'setFnExample':

> setFnExample = setFn "setFnExample" fnExample

But there is a problem.  'setFn' relies on 'store' (Ivory.Language.Ref) which
has the signature:  IvoryStore a => Ref s (Stored a) -> a -> Ivory eff ()

Predictably, 'store' does not work on anything that is not in IvoryStore,
"Things that can be safely stored in references."  ProcPtr, despite being
something that (as far as I understand it) can safely be stored in a reference,
is not in IvoryStore.  (While IvoryArea a => IvoryStore (Ptr Global a), I am
unaware of this working with Def or ProcPtr and the minimal compliant type
I could construct was Ptr Global (Stored (ProcPtr proc)) which is not a
correct type for what I am doing.)

This commit simply adds ProcPtr to IvoryStore, and enables the above code to
run.