Coroutine: Ensure all Stmt types are handled.
Refactor extractLocals' recursion pattern as a right fold to eliminate a
little boilerplate. Then handle the simple statement types in the same
pattern match as the complex ones, and eliminate the wildcard pattern,
so GHC can warn if we miss any.

From this exercise I discovered that I'd missed the Assume statement,
which was a bug. I also hadn't explicitly handled Comment statements,
but they didn't need anything special.