Revert "CSE refs that are used both const and non-const."
This reverts commit 8f3f3925b35708dd834ff9f5de269d67a85da9b0.

That patch didn't work when given an expression with multiple layers of
indirection, such as `n_var0->orient[0]`. It produced code like this:

    const float* n_cse6 = n_var0->orient;
    float* n_cse8 = &n_cse6[(int32_t) 0];

n_cse8 was used writable, but that didn't mark n_cse6 as being writable
in turn, so CSE defaulted to n_cse6 being read-only. The resulting C was
invalid.